Use built in cake tool for calling java/liquibase
Resolving the tool did not work properly on linux so
switching to the official way of calling tools in Cake
seems to be the best way forward.
